🚑 COUNT 필드 타입을 Long으로 변경
- JPQL COUNT 결과를 안전하게 매핑하도록 long → Long으로 수정

+++ src/main/java/knu/team1/be/boost/task/dto/TaskMemberSectionResponseDto.java
@@ -36,12 +36,12 @@ public record TaskMemberSectionResponseDto(
         Member member,
         List<Task> tasks,
         int limit,
-        Map<UUID, Integer> fileCountMap,
-        Map<UUID, Integer> commentCountMap
+        Map<UUID, Long> fileCountMap,
+        Map<UUID, Long> commentCountMap
     ) {
-        Map<UUID, Integer> safeFileCountMap =
+        Map<UUID, Long> safeFileCountMap =
             fileCountMap == null ? Collections.emptyMap() : fileCountMap;
-        Map<UUID, Integer> safeCommentCountMap =
+        Map<UUID, Long> safeCommentCountMap =
             commentCountMap == null ? Collections.emptyMap() : commentCountMap;
 
         boolean hasNext = tasks.size() > limit;
@@ -50,8 +50,8 @@ public record TaskMemberSectionResponseDto(
         List<TaskResponseDto> taskResponseDtos = tasks.stream()
             .limit(limit)
             .map(task -> {
-                int fileCount = safeFileCountMap.getOrDefault(task.getId(), 0);
-                int commentCount = safeCommentCountMap.getOrDefault(task.getId(), 0);
+                int fileCount = safeFileCountMap.getOrDefault(task.getId(), 0L).intValue();
+                int commentCount = safeCommentCountMap.getOrDefault(task.getId(), 0L).intValue();
                 return TaskResponseDto.from(task, fileCount, commentCount);
             })
             .toList();
